Move unit tests into the tests/unit directory
